Add check for the last reason thrown.
When an unhandled, non-promise error gets thrown, the top-level
error handler ends up throwing in a mostly-endless loop: once
for the error itself, and then once when rethrown by the promise
`.then` on the next cycle of the run loop.

Prevent cycling through these rethrows by tracking the last
reason something threw for each label, and if the reason is the
same as a previously thrown reason, simply return, and unset the
reason so that if the same error is *actually* thrown again later,
it won't be swallowed.